net: Include bond hwaddr in KernelConfig report
Bond hwaddr is persisted in the unified configuration but is missing
from the kernel configuration report. This caused several functional
tests to fail and on restoration the diff between the persistent and
kernel config would falsely detect a change.

The hwaddr attribute is now reported in KernelConfig if it was mentioned
in the unified configuration.

Change-Id: Ia67d742ba76c28ff52f5f2df02f66ec329694f1b
Signed-off-by: Edward Haas <edwardh@redhat.com>