volume: Return volume type in getVmVolumeInfo
When we prepare an image, we add information about the leaf volume in
the response. We used to return {"volType": "path"} for both file and
block volumes, and later in the virt Drive class, try to check the
volume type using os.stat(). Calling os.stat() on NFS storage may block
for unlimited time if the storage become non-responsive, causing the
entire process to be in D state.

Now we return {"type": "block"} for block volumes, and {"type": "file"}
for file volumes, eliminating the risky os.stat() call in virt layer.

Change-Id: I7cba4a47970970b67dd298991259203dfd853e06
Bug-Url: https://bugzilla.redhat.com/1530072
Signed-off-by: Nir Soffer <nsoffer@redhat.com>