pylint: Silence pylint warnings for libvirt events
pylint cannot tell that unpacking the *args tuple is safe, but we know
that the unpacking is safe since we assume libvirt documneted callback
api.

The real solution to this issue is to create a method per callback with
fixed number of arguments, and register the callbacks separately with
libvirt. But this requires hard work, untangling the way we create
libvirt connection for handling events.

Change-Id: I65df631bfac0ef17399ee9e19689c61cd6a8fd86
Signed-off-by: Nir Soffer <nsoffer@redhat.com>