virt: graphics: displayNetwork may be absent
The 'displayNetwork' parameter may be missing from the
graphics device spec parameters. This most often happens
with Hosted Engine.

The Graphics device code has already some code to deal with 'None'
displayNetwork, but this is one half-baked solutionm byproduct
of some past backward compatibility code which is no longer relevant.

This patch makes the fallback code fully baked, dealing with missing
displayNetwork.
It is not clear, and should be addressed in future patches, if
as per API contract the clients are actually at fault, and they
should always send displayNetwork.
For the short term, this patch accepts what existing Hosted Engine
deployments are already using.

Change-Id: Ib37d45e7292c7d99c63cf403e15b2f3149161baf
Related-To: https://bugzilla.redhat.com/1524119
Signed-off-by: Francesco Romani <fromani@redhat.com>