vm: graphics: don't inject displayNetwork=None
We have code in the Vm class to preserve backward compatibility
with Engines <= 3.5.

In circa-3.5 times, displayNetwork was a top-level vm.conf key.
Around the 3.6 cycle we moved it into the graphics device spec parameters.
For backward compatibility, we copied there from vm.conf if missing.
One unexpected side effect is that we injected

  'displayNetwork': None

in the graphics device specParams.

We should avoid this, because it could hide other bugs, like it
happened in rhbz#1524119

Change-Id: I483668602569a326867a9ebbbb734ef265971630
Related-To: https://bugzilla.redhat.com/1524119
Signed-off-by: Francesco Romani <fromani@redhat.com>