virt: Handle missing network parameter in Interface.get_metadata
When a SR-IOV passthrough network device is hot plugged, `network' parameter is
not set and may not be present in Interface instance.  However when the hot
plugged device metadata is updated it tries to access Interface.network
attribute, which results in AttributeError and incomplete Interface instance
initialization after successful device hot plug, since
Interface.update_device_info is not called due to the error.  Consequent
problems such as inability to handle the hot plugged network device or
tracebacks in periodic functions accessing the device instance follow.

This patch fixes the problem by checking whether `network' attribute is present
and not accessing it and not writing <network> to metadata if it is not.

Change-Id: I986aecd7fd81b4f2dbc89cddf9eb4411eac340bb
Bug-Url: https://bugzilla.redhat.com/1526133
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>