dpdk: check if 'driver' exists in device configuration
We check if dpdk supported devices exist using `lshw -json` output,
in this case, we are checking if device['configuration']['driver']
is one of the supported drivers.
Previously cases where NICs were not bound to any driver were not supported
(raised an exception), now we are also considering these cases.

Change-Id: Ibf533d433c659cc86affc12c19b236f26a5ec895
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1526192