net: DefaultRoute move semantic
VDSM supports a single network which has DefaultRoute set.

This patch changes the semantics of setting the default route from
defensively validating it is set on a single network to an active move
operation. In case a setup network is set with the default route and an
already existing network has it defined, the canonicalize step will
auto-generate the request to remove the default route from the existing
network.

Please note that a move operation implicitly generate the setup request
on the existing network, causing it to be removed and added.

Change-Id: I527c06e47bb0657987e656f946a5b032f028f785
Bug-Url: https://bugzilla.redhat.com/1522971
Signed-off-by: Edward Haas <edwardh@redhat.com>