virt: Catch NotConnectedError in guest mapping metadata sync
When changed disk mapping is detected on stats retrieval, domain metadata is
updated accordingly.  Contingent libvirtError on metadata update is caught and
logged and disk mapping metadata is supposed to get updated in one of future
stats calls.

However libvirtError is not the only error that can prevent metadata update.
It won't work also when the domain is not connected, such as during VM start or
when the VM is finished.  NotConnectedError is raised in such a case and whole
Vm.getStats call fails due to the error.  So NotConnectedError should be caught
and handled at that place too.

Change-Id: Iabfc402b6a8cee0a84c20c84a15afd8ad7b86047
Bug-Url: https://bugzilla.redhat.com/1525955
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>