virt: Don't update disk mapping when VM is not Up
Updating disk mapping when a VM is not in UP status doesn't make much sense and
may be harmful.  For instance, guest agent data is not available anyway before
the VM is started and we don't care about handling disk mapping updates when a
VM is already down.  And when a VM is migrating, we must even avoid disk
mapping metadata updates, since libvirt blocks on setMetadata until the
migration finishes.

So let's update disk mapping only when the VM is not UP.  In case disk mapping
change is detected and ignored in such a case, while the update is still
meaningful, the mapping will be updated once the VM switches to UP status
again.

Change-Id: If19db4ffed959a516b4282032a497a563779199c
Bug-Url: https://bugzilla.redhat.com/1527416
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>