virt: Return None disk mapping hash when disk mapping is empty
Vm class updates and processes disk mapping report from GuestAgent.  However
guest agent is not available initially or it may not be available at all.
GuestAgent returns empty disk mapping in such a case but it's still being
processed and attempted to store to metadata.  That's useless and can be even
harmful, e.g. when we try to update metadata before Vm domain is available and
connected.  To avoid such a processing we must check that disk mapping is
empty.  We already check disk mapping hash in Vm._update_guest_disk_mapping so
if a special hash value, namely None, was returned from GuestAgent on empty
disk mapping, the processing would be skipped automatically.

This patch changes GuestAgent.guestDiskMapping to return None on empty disk
mapping in order to skip the useless and possibly harmful disk mapping
processing as described above.

Change-Id: Icfcd3f3e991b2b21de3a30f753c4d0637885ce74
Bug-Url: https://bugzilla.redhat.com/1525955
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>