net: Add OVS net gateway normalization in kernelconfig
kernelconfig.normalize is used to normalize a given running config in
order to compare it with kernel config.

With an OVS network, a gateway which is not a default route is
not supported (source route is not supported).

This patch adds a new normalization for the OVS networks gateway
property to remove gateway when no default route is specified.

Change-Id: Ifbc6daf23393f4d70884513cbf37bc12bf23b2a2
Signed-off-by: Edward Haas <edwardh@redhat.com>