virt: Always remove unplugged network device from Vm.conf
When we hot unplug a network device specified by XML, we don't remove it from
Vm.conf assuming that it's not needed with XML Engine.  However we still have
to keep legacy conf data around, e.g. for migrations to older Vdsm versions.

This patch ensures that Vm.conf is updated in any case.

Change-Id: Id896a65cc87196b0954d6d4a4ee803f1ee72401b
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>