virt: Fix swapped variable value in Vm._update_guest_disk_mapping
At the beginning of the method the guest agent disk mapping hash should be
stored to a local variable, to be later used as new Vm._last_disk_mapping_hash
value.  This is to prevent races with disk mapping updates in the meantime, so
that we don't miss any update.

However, Vm._last_disk_mapping_hash was stored to the local variable by
mistake.  That means Vm._last_disk_mapping_hash has never been updated and disk
mapping metadata being updated on each call.  This patch fixes the mistake.

Change-Id: I91dadd4a478d2e4438490355299499ba25fbba62
Bug-Url: https://bugzilla.redhat.com/1527416
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>