Create lockfile.lock func  to avoid running multiply instances of vdsmd
To avoid running executable twice we use a lockfile. The lockfile module
contains lock function that can be use when starting the executable.
First instance can get the lock file, second try throws exception and
exit.
On D state we can restart the service and depends on the system, we might
get to a state that the daemon controller runs two instances of the same
service executable. this lock mechanism helps to avoid such cases.

Change-Id: I6f13409d156a080969ef63f6829337121d66ba27
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1525171
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>