core: add 'shared' field to METADATA_KEYS
Since compatibility version 4.2 the engine composes the VM XML that
passed to libvirt.
As part of this change, the XML should contain a field to indicate
whether the VM disk device is a transient (backup) disk or not.
If the disk is transient, the VDSM should create a transient volume
under /var/lib/vdsm/transient on the relevant host.

In the engine - missing 'shared' field was added to VM disk device metadata
in the domain XML.

 <metadata>
    <ns0:qos />
    <ovirt-vm:vm>
        <minGuaranteedMemoryMb type="int">1024</minGuaranteedMemoryMb>
        <clusterVersion>4.2</clusterVersion>
        <ovirt-vm:custom />
        <ovirt-vm:device devtype="disk" name="sda">
            <ovirt-vm:shared>transient</ovirt-vm:shared>
            <ovirt-vm:imageID>...</ovirt-vm:imageID>
            <ovirt-vm:poolID>...</ovirt-vm:poolID>
            <ovirt-vm:volumeID>...</ovirt-vm:volumeID>
            <ovirt-vm:domainID>...</ovirt-vm:domainID>
        </ovirt-vm:device>
        <launchPaused>false</launchPaused>
        <resumeBehavior>auto_resume</resumeBehavior>
    </ovirt-vm:vm>
</metadata>

Added the missing 'shared' field to METADATA_KEYS in order to parse it
from the domain XML.

Change-Id: Icbe73addfff8cba6159d127bfcbc78497f66de62
Bug-Url: https://bugzilla.redhat.com/1526815
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>