vm: Do not check disk type when starting LSM
If engine did not specify the disk type, we used to check the disk type
using utls.isBlockDevice, which may block for minutes if the disk is on
a non-responsive NFS storage domain. This check is not needed now since
prepareVolumePath adds the disk type if needed.

Remove the disk type check, and update the disk replica only if disk
type was added or modified by prepareVolumePath.

Change-Id: I06c0ec7073898282ce37e195fbc05846a6553b4e
Bug-Url: https://bugzilla.redhat.com/1530072
Signed-off-by: Nir Soffer <nsoffer@redhat.com>