lvm: Do not refresh lvs static lvs
Since change I845ce1f930d we are refreshing active LVs when activating
LVs. This is required for lvs which may be extended to prevent data
corruption if the lv size on the host is not correct.

However for static lvs like most of the special lvs, this refresh is not
needed, and causes additional load on the storage server for no value,
and lot of noise in vdsm logs.

lvm.activateLVs has now a ``refresh`` argument, allowing disabling
refresh for LVs which do not need a refresh. We use the new option to
avoid refreshing static LVs.

With this change, only the metadata lv is refreshed when activated.

Change-Id: I7118099340e2681b5d07301abd75dab39ea41f56
Bug-Url: https://bugzilla.redhat.com/1432039
Signed-off-by: Nir Soffer <nsoffer@redhat.com>