host stats net: Refactor network statistics gathering
Network statistics report is now exposed through the network.api.
It gathers the link statistics using the link package.

The stats are no longer collected using the sampling mechanism, as there
is no need to do so anymore. On each getStats request, the network stats
are being collected directly.

In the following patch, the network stats is going to be exposed as an
independent verb (VDSM API).

Change-Id: Id26c543e7b560c93bdd2f26766c7e2bb24086291
Signed-off-by: Edward Haas <edwardh@redhat.com>