virt: Set SourceThread._started flag for hibernation
SourceThread._started flag was introduced in 8469d6b.  However it is set to
True only for host migrations, not for hibernation.  The result is that
when hibernation succeeds the loop in SourceThread.run continues, fails on a no
longer existent domain and reports confusing errors to the log.

This patch fixes the problem by setting SourceThread._started flag also for
file migrations.

Change-Id: I5a95e7d27442572dfbb741356a7491f59e90bae1
Bug-Url: https://bugzilla.redhat.com/1530292
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>