hooks net: Remove the fakesriov hook
With the introduction of a dedicated network caps, and the will to allow
Engine to use it, either a new hook point should be created or drop any
existing hooks that manipulate the network caps data.

It has been decided to drop the network caps for now, especially when
the existing one is not used in a production env.
Therefore, this patch removes the fakesriov hook.

Change-Id: I93b4fa69f502bd3f23f3f6fd28eabd9d5f23e946
Signed-off-by: Edward Haas <edwardh@redhat.com>