tests: Speed up concurrent.format_traceback test
We use 1 second sleep for giving time for the worker thread to start and
enter builtin time.sleep(). This works but make the test slow, so it
does not run in the CI.

Replace the sleep with 2 events - this ensures that we format a
traceback when the worker thread is inside the inner() function, without
sleeping.

To test the formatted traceback, we don't try to check the function the
thread is blocked on - this is racy and require sleeps. Instead, we
check that the both the worker() and inner() function are mentioned in
the traceback.

This makes the test both deterministic and 1000 times faster so we can
always run it.

If a slave is horribly slow and the worker thread failed to start in 1
second, we will fail with RuntimeError, making it clear that this is not
a test failure but an issue with the environment.

Example run:

$ ./run_tests_local.sh common/concurrent_test.py:TestFormatTraceback.test_get_running_trace
common.concurrent_test.TestFormatTraceback
    test_get_running_trace                                      OK

----------------------------------------------------------------------
Ran 1 test in 0.001s

Change-Id: I2765c31729821fa31f98ebf88a40d7e129534cbc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>