net: Net restoration should not remove unowned bonds
When the network restoration detects differences between actual and
persisted (desired) state, an attempt to sync the two is made.
Specifically, when external bonds exist, the restoration process will
wrongly attempt to remove them (as they do not appear in the persisted
state).

Such external bonds which do no appear in the persisted state will no
longer be removed in the network restoration process.

Change-Id: Ieaf184c3d8ed52c885d52942b8347b710fbbd565
Bug-Url: https://bugzilla.redhat.com/1530839
Signed-off-by: Edward Haas <edwardh@redhat.com>