vm: support error_policy='report' for CDROMs
Since patch I62ff73048cd5bef59a7c3236fcc8c91b0254a660
Engine wants to set error_policy='report' for CDROM devices.
To do so, it sets the right property in the XML it creates and
sends to Vdsm.

Problem is Vdsm doesn't (yet) pass through that XML, but
it parses that data, initialize the VM data structures, and
produce a new XML from that, which should be equivalent
to the one it received.

It is not the case, because Vdsm did not create the 'driver'
element for CDROM devices alltogether.

This patch fixes the issue making the Vdsm code recognize
the error_policy = 'report' and producing the right XML
in this case.

Related-To: https://bugzilla.redhat.com/1207992
Change-Id: I850d279f24c3f2ad401f7f93dd31bc29abdaa873
Signed-off-by: Francesco Romani <fromani@redhat.com>