net: Do not preserve ifcfg IP data for bond and bridge slaves
During the ifcfg configuration, in some scenarios the defined IP
configuration in the existing ifcfg file is preserved.
(When a base device is not owned by VDSM, but a VLAN on top of it is)

In this patch, IP configuration is not loaded from the ifcfg existing
file if the device in question has (or needs to have) a bridge or
a bond over it.

Change-Id: I005ea25176f130092e5c1d25dbe945cb986f37d5
Bug-Url: https://bugzilla.redhat.com/1530839
Signed-off-by: Edward Haas <edwardh@redhat.com>