net: Import valid IP data from an ifcfg file
When the IP configuration is retrieved from an existing ifcfg file, the
static IP configuration is valid only if all values exist.
When NM is active, ifcfg file may contain partial configuration values
which may not be used with a pure ifcfg setup.

In case one or more values are missing, all IP values are considered as
missing.

Change-Id: I881897973ffa382e98e0b24616ba5736b5590915
Bug-Url: https://bugzilla.redhat.com/1530839
Signed-off-by: Edward Haas <edwardh@redhat.com>