virt: Set correct disk paths on XML Engine resume
When a VM is resumed from a snapshot, the disks in its domain XML must be
updated to the drives obtained from Engine.  In the legacy non-XML Engine path
it was ensured by initializing disk devices from conf provided by Engine and
replacing the disk attributes in domain XML with that information.  However
with XML Engine the only source of information used is the domain XML from the
snapshot, which is wrong: Devices are initialized from it and the domain XML is
updated with information from itself, i.e. not updated at all.

To fix that both Engine and Vdsm patches are needed.  Engine must provide
previously omitted `xml' parameter on resume from snapshot, with up to date
drive parameters.  And Vdsm must take the updated disk properties from that
parameter.  Only disk attributes are updated, the rest of the domain XML must
be used from the snapshot untouched.

This patch implements the Vdsm changes.  We must ensure that all occurrences of
XML data are properly updated and the updated data is used everywhere, so we
perform the update as soon as possible, that means in Vm.__init__.

Moreover there are the following situations we must distinguish and handle
properly:

- Modern way, `xml' param is present: We use _srcDomXML and update it with that
  param.
- Legacy way, no `xml' param, but legacy configuration is present: We use
  _srcDomXML and update it with devices created from conf provided by Engine.
- Intermediate way, no `xml' param and no legacy configuration (Engine 4.2.0):
  This is wrong and doesn't provide necessary configuration, but we can't crash
  on that; _srcDomXML is used unchanged.

Change-Id: If3ccbfd86b086b454a1fcc52f2bfbf086c3962c3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>