tests: periodic: add test for exclusive operation
Due to a bug in the implementation of exclusive periodic operations,
if we try to schedule while the executor is full, the attempt fails
(ResourceExhausted), and it is never tried again.

This is wrong: the semantic of the exclusive operations is that
at any given time exactly one given operation should be either
1. executing (including blocked)
2. scheduled to be executed

So the correct behaviour is to schedule again the operation,
trying until we succeed to submit it to the executor.
Once we succesfully submitted the operation to the executor,
we should schedule it again once it completed -somehow.
The latter part is already implemented.

Add test to demonstrate the bug.

Notes about backporting to 4.1
- exceptions.ResourceExhausted does not exists, using
  the old executor.TooManyTasks instead.

Change-Id: Ic787e0979324b89bf141c1c0fde5752356b6a029
Bug-Url: https://bugzilla.redhat.com/1512547
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>