virt: periodic: reschedule exclusive operations
The semantic of exclusive operations is that no more than
one operation should be running at any given time.
This means that one periodic exclusive operation is always either:
1. running (this includes being blocked)
2. waiting to be runned, thus scheduled and pending execution

This means that we should schedule again one new exclusive operations:
1. once it is completed. This is because of the semantic explained just
   above.
2. iff we failed to submit to the executor. This is because
   the executor never loses data: once one operation is submitted,
   it *will* be executed (sometime in the future, asap)

Due to one oversight, #2 was missing, and this patch implements the
logic.

Change-Id: Ib2bc51abf0b6726646999998d68ed36b3499f500
Bug-Url: https://bugzilla.redhat.com/1512547
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>