net: NM device needs to be stateless
Currently, the NM device object keeps the device in its state. Doing so
assumes that the device will not disapear before calling other actions
on the object. Such an assumption has been found unsafe, as the device
on which the object is based can be removed by NM in mid-air.

With this change, the device is no longer saved in the object with
re-fetched on each operation.
If the device no longer exists, the NMDeviceNotFoundError exception is
raised.

Change-Id: I3f27b7c8e008150a2bca49bde91b579a5fed6e4e
Bug-Url: https://bugzilla.redhat.com/1538632
Signed-off-by: Edward Haas <edwardh@redhat.com>