tests: Extract qemuio module
This module provides the qemu_pattern_write and qemu_pattern_verify,
used to write data to qcow2 files and verify qcow2 files content.

Moving the functions to separate module will allow porting qemuimg tests
to python 3. Currently they cannot run on python 3 since storagetestlib
imports modules which are not compatible with python 3 yet.

The tests for the helpers moved to a new module and converted to pytest.

Change-Id: I8a1e747c9b98e61c74086664aa72b3db542a6c68
Signed-off-by: Nir Soffer <nsoffer@redhat.com>