spec: require kernel for safe use of blkdiscard
BZ#1507043 describes a case where using blkdiscard on a storage array
that misreports its limits may result in an illegal request. This
problem was fixed by kernel-3.10.0-693.14.1.el7 that's available for
both RHEL and CentOS 7.4.z updates. This patch requires that kernel
version so that blkdiscard can safely be used with those arrays.

Change-Id: I527259e6340d6f9d1469e0131d55df8d5086c5a3
Signed-off-by: Allon Mureinik <amureini@redhat.com>