vm: introduce ReplicationNotInProgress exception
Currently, in diskReplicateFinish, when there is
no replication in progress or there is an error in
the replication process, the same exception will
be sent to the engine.

However, when the diskReplicate attribute is not
available as a result of some failure we should be
able to perform a cleanup to make the next LSM
possible without a manual cleanup.

Change-Id: I776a709f6eaffe387a08c47acb3b7d9bbded540a
Bug-Url: https://bugzilla.redhat.com/1542423
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>