multipath: Do not fail I/O after short outage
Since 3.6 we use "no_path_retry fail" for all devices by default.
This setting is not new, but before 3.6 it was applied only to a few
devices specified in our multipath conf, while other devices used the
defaults hard-coded in the multipathd daemon.

We have seen several reports in the user list showing that this setting
makes oVirt too fragile considering short storage outage, and causing
vms to pause too quickly.

This patch changes the default setting to "no_path_retry 4". With this
setting, once multipathd notices that the last path has failed, it will
check all paths 4 more times (20 seconds assuming the default 5 second
polling_interval). If no path is up, it will tell the kernel to stop
queuing.  After that, all outstanding and future I/O will immediately be
failed, until a path is restored. Once a path is restored the 20 second
delay is reset for the next time all paths fail.

This change introduces a 20 seconds delay after the last path has
failed, before outstanding I/O will fail. If there is a short storage
outage on the server, and the server could recover in 20 seconds, no vm
will pause, and commands ran by vdsm (e.g. lvm) will not fail.

Note that this patch does not solve the issue of vm not resumed after a
short storage outage, but it can decrease the chance that a vm will
pause in this case.

This change was merged in I6496dbdaafca6b110c952fcc5d51bf9ac04d49b4 but
had to be reverted because of multipath regression in EL 7.3. The
device-mapper-multipath version we require includes the fix for this
regression.

Change-Id: I5455df7a89ff8f7ea5123334a91c6ebfe4e5c552
Bug-Url:https://bugzilla.redhat.com/1335176
Signed-off-by: Nir Soffer <nsoffer@redhat.com>