virt: metadata: save deviceId if available
We need to save deviceId in the metadata, to be able to fully
rebuild a faithful device state on recovery. Otherwise, Engine
will not be able to match the devices.

We will have this attribute set (not None) only if the VM was started by
a 4.1 entity, so we can safely attempt to save it every time: for 4.2
VMs, we will silently skip it

Change-Id: I4767313aaaec181550510078b89895f8e3630ddd
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>