vm: metadata: try to use backward compat params
To ensure full 4.1 compatibility, Vdsm needs to report the deviceId
of the devices of a VM. This requires a Device is given its deviceId
in the first place when built out of the XML.
Vdsm 4.2 always recover from XML, regardless the cluster it is installed
on (4.1 vs 4.2 vs 4.y), so, if the deviceId is given, let's use it.

This patch makes the device always use this field even if unneeded,
e.g 4.1. cluster later upgraded to 4.2. If Engine doesn't need this
field anymore, the only real drawback is wasted space in the domain
metadata.

Change-Id: I571a4f5b339a8e5ccd4cfb20f563c2a897c5d1d1
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>