virt: vmdevices: fix special handling of config()
In commit 2b98daffb2c2dec we made the config() method
of some device classes return None.

The config() method is called in the backward-compatibility
flow, when Vdsm has to rebuild the vm.conf to work in a 4.1
environment, to return meaningful values in Vm.status().
In the compatibility flow, if calling config() on a given device
instance return None, nothing is added on vm.conf['devices']
for that device.

The rationale to return None for some device classes, is to make sure
vm.conf['devices'] is updated once: either before the Vm is started,
code added in 4.2, or after the Vm was started, in the existing
Vm.getUnderlyingDevicesInfo() method.
Doing this lead to simpler and safer code: we avoid entirely
the problem of recognizing and updating devices, adding duplicates
and so forth.

However, this is not good enough when Vdsm 4.2 runs in 4.1 cluster:
until the Vm is running, any call to Vm.status() will return partial
vm.conf. Furthermore, the code in Vm.getUnderlyingInfo() *will*
add the missing devices, but with minimal information.

We have another approach, implemented in this patch:
Let's just leave vm.conf alone completely in the Vm.run(), and
rebuild the device configuration data on the fly in Vm.status().

This approach keeps the advantage of having just one piece of code
which changes vm.conf['devices'], and make it possible to
fully rebuild the device configuration avoiding any changes to
the existing code, thus maximizing backward compatibility.

Change-Id: Ibd5e92741c7b1f5c74767e9647a94f680a50f7c4
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>