vm: export memGuaranteedSize in status
We need to report memGuaranteedSize to Engine when
running in 4.1 cluster.
Vdsm already takes care of all the bookkeeping, the
only missing step is to add the value to the response
of getVMFullList, so in the return value of Vm.status().

To do so, we must also make sure to restore the value
Vdsm saved, but only if we Engine didn't set itself.

Change-Id: Ia9b270d4e401f732deda23f56208f571db0512de
Backport-To: 4.2
Related-To: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>