openstack_net hook: report neutron names of host
Engine needs to know how Neutron referes to the host. This new hook
reports it in getCapabilities.

With this patch, vdsm now reports a 'openstack_binding_host_id'
dictionary, mapping a pluginType to the host name used by the plugin.
This suggested api allows for future assignement of a host to more than
a single external network provider.

Bug-Url: Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1554268
Change-Id: If03b57687066866367d8084734944a497079e2c5
Signed-off-by: Dan Kenigsberg <danken@redhat.com>