vm: migration: don't add conf['xml'] if src is 4.1
When initializing a VM, we put the _srcDomXML content in
the 'xml' field, and we use it as authoritative source of
truth for the domain definition.

And this is fine if we do that once the original XML is in
turn authoritative and complete, and this is true only
for 4.2 sources, being those Engine calls or migrations from
4.2 Vdsms.

When we migrate from 4.1 sources, we should not consider this
XML truthful until libvirt gives it back to us, and pre-initialize
Vdsm with data from vm.conf.

Once the domain is started, we reconcile the data Vdsm received
to the XML from libvirt, and from that moment on it doesn't matter
anymore the source of the XML - the XML is 4.2 compliant, otherwise
we will have other bugs, for example in recovery.

Change-Id: I86f2d78837c91b97afe11233bec74daf03466121
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>