virt: vmdevices: fix attribute reporting
Careful comparison of the data we get from XML and from the vm.conf
reveal that we misreport few fields.
This patch makes the XML path behave like the vm.conf path.

Please note that the meaning of Device classes fields 'type', 'device'
and 'deviceType' is highly irregular, and so is the libvirt
representation. There is no one real correct mapping, so we need
to preserve backward compatibility with what we did in the past.

Change-Id: I294f7ee4d477dd49eff79185147779ba16d3cbf9
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>