virt: metadata: add set_custom helper
In 4.2+ clusters, Vdsm is not expected to add custom variables
in the metadata. Those variables are expected to be set by Engine,
consumed by Vdsm and and never changed by the latter.

Things change if Vdsm 4.2 is run in 4.1 clusters. Now we use
the domain XML as source of truth for recovery, so we need to
add the custom variables that we previously stored in the
recovery files, otherwise we will lose them on restart.

This patch adds support to inject custom properties,
the tests, and let the Vm class use this new facility.

Please note that even though looks like Engine doesn't care about
receiving back or not the custom properties it set, Vdsm
needs them to properly initialize itself in some cases, mostly
related to network area (e.g. NIC devices)

Change-Id: Ic6600ae9dc403ed720c6fc47e404f24843d27285
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>