vmdevices: fix config() in backward-compatible way
In 4.1 compatibility mode, we cannot expose the devices' configuration
as Vdsm stored it, because in the initialization flow
some normalization/trasformation has been done.

Furthermore, between 3.6 and 4.1, the API got saner, and we now send
more meaningful values, like true booleans instead of True,
or less convoluted way to express the configuration data (e.g cdrom
paths).

This patch starts the compatibility layer, implementing the
reverse of the normalization Vdsm does for backward compatibility sake.
We isolate those changes in one vmdevices.compat module, to make
it easier to kill it as soon as we can.

At the moment, we don't aim for full backward compatibility, which
is not needed anyway. Most of the vm.conf is meant to be set by Engine
and consumed by Vdsm; when checking the Vdsm responses, Engine reads
only few selected fields, most notably 'deviceId', 'alias', 'address',
for VFIO NICs, 'hostdev', for disks, 'volumeChain'.
All of those are already returned by Vdsm even in the backward
compatible flow.

We used to dump all the vm.conf in the responses to Engine because
it was simpler, but it always was a bad idea.

Change-Id: I179c40d67db2e2f4c098c11e535a865b0cb2f7a4
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1542117
Signed-off-by: Francesco Romani <fromani@redhat.com>