net: Introduce NM device sync operations
NM device object now includes a sync-operations object, which provides
sync related operations.
The methods introduced with this change wait for a device state to be
reached and only then returns.
In case the state is not reached by a defined timeout, an exception is
raised.

On some platforms, some incidents have been detected where the device
state is not "activated", causing several properties of a device not to
exist yet.
Therefore, when a NM device object is created, it is not assumed to be
in its stable state, and where needed, a waitfor sync operations is
issued to block until the relevant state is reached.

Change-Id: Ia98f6695585a2105c54776c24768926f9d80c9dd
Signed-off-by: Edward Haas <edwardh@redhat.com>