logutils: Adaptive log level
Add adaptive mode to ThreadedHandler. In this mode, we try to drop lower
priority messages faster, to avoid filling the queue.

When queue is 60% full, we start to drop DEBUG messages. This
significantly decreases the number of logged messages, so in most cases
no more message will be dropped. If decreasing the effective log level
does not help, we will drop INFO messages, and if queue size continue to
increase, drop WARN messages, and so on.

Change-Id: I7b845e7b6d8d0bb7f7b7b1b9c63fa28badfecf14
Bug-Url: https://bugzilla.redhat.com/1516831
Signed-off-by: Nir Soffer <nsoffer@redhat.com>