vm: hotplug disk: always add disk params
We need this to properly support 4.1 backward compatibility.
The problem is that even in the XML-based, 4.2 flow, Vdsm needs
to recreate the vm.conf['devices'] data to support some storage
flows, most notably live merge. At the moment, those flow are
not yet update to take advantage of the domain XML.

Without this patch, hotplugged disk will not add their conf
data to vm.conf, thus is the operator tries to do snapshot and
then live merge on it, it will fail with no good reason and
with confusing errors.

Change-Id: I5f8dd29b20aa8bd48389a9cb1710d38051bf49d4
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Branch-Name: vdsm42-in-cluster41
Topic: vdsm42-in-cluster41