virt: metadata: add from/to tree methods
In the restore fram snapshot flow, Vdsm needs to make
changes to the source domain XML (_srcDomXML), before
to use the data to initialize itself.

The operation which needs to be done are little different
to what we already support, so a sensible choice seems
to extend the current libraries instead of re-implement
those operations from scratch, again, on top of simple
primitives like the ones provided by the vmxml.py module.

We start the extensions adding facilities to the
metadata.Descriptor class to (de)serialize itself to
ElementTree object(s), in order to better integrate
with DomainDescriptor and with the new flow.

This is a simple extraction of code we already have,
covered by existing tests.

Change-Id: If743f63b939621d525b01e9240db5b5f8265545b
Bug-Url: https://bugzilla.redhat.com/1533155
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>