virt: dom_desc: add method to get metadata desc
In the restore from snapshot flow, we need to amend
the source domain XML (_srcDomXML) we use with
fresher data received by Engine.
We use DomainDescriptor for this purpose, but we
will need to change the metadata too.

A simple solution is to keep the DomainDescriptor
own the XML document to change, and to add support to
get one metadata.Descriptor from it.

This way, any change to metadata will be easily reflected
in the XML that the DomainDescriptor owns.

Change-Id: I64db6466fb53092096215146a6afc584e0ad67e4
Bug-Url: https://bugzilla.redhat.com/1533155
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>