vm: split correctDiskVolumes
We need to change the metadata of the source domain XML
(_srcDomXML) in the restore from snapshot flow, but we
need to do that only if restoring from a snapshot created
by 4.2.

This change makes the two support flows -vm.conf based
and XML based- too different from each other, so this
patch splits the Vm._correctDiskVolumes in two,
and fixes the from_xml variant to change the metadata as well.

Change-Id: I69355af07ffd34bfe59b73d64e3d18669704b527
Bug-Url: https://bugzilla.redhat.com/1533155
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>