vm: support setting device alias for drives and nics
The engine is expected to set the alias property of devices in
the domain XML as from 4.2. It would work with all the non-storage
and non-network devices out-of-the-box since the XML is passed
through to libvirt, but some adjustment is needed for it to work
for storage and network devices.

The new alias property that the engine would send looks as follow:
<alias name="ua-<device_id>">
where device_id is the ID of the device on the engine side.

Change-Id: Iee33a9bfa2ddf2d8deac26f251082957ac022848
Bug-Url: https://bugzilla.redhat.com/1544721
Signed-off-by: Arik Hadas <ahadas@redhat.com>